# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations/{conversationId}/messages:
    get:
      operationId: listConversationMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, creator, company, recruiter, influencer, moderator, admin]
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        experience:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
        socialLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialLink'
        createdAt:
          type: string
...
    Message:
      type: object
      required: [id, conversationId, senderId, sender, content, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        isRead:
          type: boolean
        isEdited:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string

    # ─── NOTIFICATION SCHEMAS ─────────────────────────────────────────────────
...
    Skill:
      type: object
      required: [id, skill, level]
      properties:
        id:
          type: string
        skill:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
...
    Experience:
      type: object
      required: [id, company, title]
      properties:
        id:
          type: string
        company:
          type: string
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        current:
          type: boolean
...
    Education:
      type: object
      required: [id, institution, degree]
      properties:
        id:
          type: string
        institution:
          type: string
        degree:
          type: string
        field:
          type: string
        startYear:
          type: string
        endYear:
          type: string
        current:
          type: boolean
...
    Language:
      type: object
      required: [id, language, proficiency]
      properties:
        id:
          type: string
        language:
          type: string
        proficiency:
          type: string
          enum: [basic, conversational, fluent, native]
...
    SocialLink:
      type: object
      required: [id, platform, url]
      properties:
        id:
          type: string
        platform:
          type: string
        url:
          type: string
...